Fixes D3D11 Renderer to work on 64 bits builds

+++ src/Renderer/D3D11/D3D11GraphicsDevice.cpp
@@ -276,7 +276,7 @@ bool D3D11GraphicsDevice::CompileShaderFromFile(const char* inFileName, const ch
 	//now copy to vector if we like it...
 	if (pBlobOut)
 	{
-		int compiledCodeSize = pBlobOut->GetBufferSize();
+		size_t compiledCodeSize = pBlobOut->GetBufferSize();
 		outCompiledShaderCode.resize(compiledCodeSize);
 		std::memcpy(outCompiledShaderCode.data(), pBlobOut->GetBufferPointer(), compiledCodeSize);
